LayoutTests/fast/message-channel/transfer-mediastreamtrack-expected.txt

Test that a MediaStreamTrack can be transferred through a MessageChannel in the same context.

PASS: received a MediaStreamTrack
PASS: track kind is audio
